This is an attemp to fix problems in building FLTK2 on x86_64 after recent changes in ShortcutAssignment. I had the problem on x86_64 Linux box, and apparently "long long" is the proper type on Windows Vista, so looks like the only way to solve this problem in a portable way is to use size_t instead of unsigned int.
git-svn-id: file:///fltk/svn/trunk@6517 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/Widget.h
@@ -22,6 +22,8 @@
 #ifndef fltk_Widget_h
 #define fltk_Widget_h
 
+#include <cstdio>
+
 #include "Style.h"
 #include "Rectangle.h"
 
@@ -101,12 +103,12 @@ public:
   const char *tooltip() const	{ return tooltip_; }
   void	tooltip(const char *t)	{ tooltip_ = t; }
 
-  unsigned shortcut() const	;
-  void	shortcut(unsigned key)	;
-  bool	add_shortcut(unsigned key);
-  bool	remove_shortcut(unsigned key);
+  size_t shortcut() const	;
+  void	shortcut(size_t key)	;
+  bool	add_shortcut(size_t key);
+  bool	remove_shortcut(size_t key);
   void  remove_shortcuts()	;
-  unsigned label_shortcut() const;
+  size_t label_shortcut() const;
   bool	test_label_shortcut() const;
   bool	test_shortcut() const	;
   bool  test_shortcut(bool) const;
